refactor: extract rig path helpers
